requirements.txt
# Web Application Framework
streamlit==1.49.1

# Medical Imaging & Data Processing
imageio==2.36.0
numpy==1.26.4
pandas==2.2.3
Pillow==11.2.1
pydicom==3.0.1
SimpleITK==2.5.2
nibabel==5.3.2

# Utilities & Visualization
matplotlib==3.9.2
tqdm==4.67.1
requests==2.32.3

# ML
transformers==4.54.1
huggingface_hub==0.34.3
accelerate==1.9.0
bitsandbytes==0.46.1

# API
fastapi==0.112.0
uvicorn[standard]==0.30.3
python-multipart==0.0.9

# torch, torchvision, and flash-attn are installed separately in Dockerfile